backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI and ML
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2

# Email and notifications
fastapi-mail==1.4.1
celery==5.3.4
redis==5.0.1

# File handling
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# CORS and security
fastapi-cors==0.0.6
python-dotenv==1.0.0
